backend/requirements.txt
fastapi
uvicorn
python-multipart
PyJWT
scikit-learn
joblib
numpy
pytesseract
Pillow
